Improved/simplified framework for frontend checkers.
Summary:public
Improved/simplified framework for fronend checkers.
Now we have a unique hook from cTrans to run checkers on statements and a unique
hook from cFrontent to run checkers on declarations.
So now when adding a checker we don't have to modify cTrans/cFrontend.

Moreover made more sistematic the way checkers are invoked. This simplify the definition
of checkers and the way we use them.

Code is now simpler.

Reviewed By: jvillard

Differential Revision: D2976589

fb-gh-sync-id: fbe22d4
shipit-source-id: fbe22d4