Revamped Timeout module to save and restore seconds and symops.
Summary:public
Revamped Timeout module by storing elapsed wallclock seconds, and the status of symops, in case of recursive calls.
Extended the API with suspend() and resume() to pause and resume the current timeout.
These are used before and after an on-demand call to the analysis functions.
This achieves the effect that each procedure, even though is interrupted, has its own time and symop counters, which are suspended and resumed as required.

Reviewed By: jeremydubreil

Differential Revision: D2976918

fb-gh-sync-id: 0ed1079
shipit-source-id: 0ed1079