Deprecate incremental
Summary:public
Deprecate the incremental mode.
Several parts of the back-end can be removed.
The options for incremental analysis -i at the python level are now deprecated, and re-routed to --reactive.
The main difference with --reactive is that it does not produce an analysis of the whole project, but is limited to what is reachable via reactive propagation starting from the changed files.

Reviewed By: sblackshear

Differential Revision: D2960078

fb-gh-sync-id: 6e8b46b
shipit-source-id: 6e8b46b