Initial support for lazy dynamic dispatch
Summary:public
Lazy dynamic dispatch handling works as follows:

Assuming a call of the form:

  foo(a);

where the static type of `a` is `A`. If during the symbolic execution, the dynamic type of the variable `a` is `B` where `B <: A`, then we create on-demand a copy `foo(B)` of `foo(A)` where all the uses of the typed parameter `a` are replaced with a parameter of type `B`. Especially, if `foo` contains virtual call, say `get` where `a` is the receiver, then the call gets redirected to the overridden method in `B`, which simulates the runtime behavior of Java.

This lazy dynamic dispatch mode is only turn on for the tracing mode for now in order to avoid conflicts with sblackshear's approach for sound dynamic dispatch.

Reviewed By: sblackshear

Differential Revision: D2888922

fb-gh-sync-id: 3250c9e
shipit-source-id: 3250c9e