Add support for @SuppressWarnings for Gradle and Ant projects.
Summary:public
Instead of using the collection of suppress warnings annotations to filter out the errors while generating the error reports, we just add this SuppressWarnings at translation time, like any other annotations, and the reporting functions in the Reporting module will just skip the errors when the method is annotated with SuppressWarnings.

This allows us to have a suppress warnings mechanism that is independant from the integration with the build system.

Reviewed By: sblackshear

Differential Revision: D3012395

fb-gh-sync-id: 35f5f9b
shipit-source-id: 35f5f9b