Add --continue option for reactive analysis to continue the capture.
Summary:public
The reactive analysis starts from the set of changed files/procedures, and proceeds
reactively to analyze their dependencies.
This means that after every command, the set of changed files/procedures is reset.
With the --continue option, the capture is continued: all the files/procedures marked
as changed stay changed, plus any additional changes are recorded.
In addition to allowing to spread capture over several commands, the option also allows to separate capture and analysis in reactive mode, or to repeat the analysis.

Reviewed By: sblackshear

Differential Revision: D3046361

fb-gh-sync-id: b6e3797
shipit-source-id: b6e3797