Refactoring in preparation of implementing C++ lambda
Summary:public
This diff refactors the current recursive module. It simplifies the structure
making CMethod_decl redoundand.

The idea is to have now two recursive functors: cTrans.ml and cFrontend.ml.
The first dealing with all the expressions and the latter dealing with all
the declarations (in a later diff, we may want to change names of these moules
to reflect Expr and Decl).

This structure will enable to implement lambdas. The previous version
would require some more complex solution where another recursive module
would be involved.

I'm breaking the refactoring in several diffs to make it easier to review.

Reviewed By: akotulski

Differential Revision: D3035122

fb-gh-sync-id: 7dabe9e
shipit-source-id: 7dabe9e