Accept command line options via environment vars
Summary:public
Generalize command line option parsing to query an environment variable
for args, and then parse the environment and command line arguments.

Each executable uses a distinct environment variable:
  - InferAnalyze:   INFER_ARGS
  - InferJava:      INFERJAVA_ARGS
  - InferClang:     INFERCLANG_ARGS
  - InferLLVM:      INFERLLVM_ARGS
  - checkCopyright: CHECKCOPYRIGHT_ARGS

For now these variables need to be set manually.  So the usability is
still pretty bad, but is a step in the direction of enabling debugging
the analyzer on code built with buck.

Reviewed By: cristianoc, martinoluca

Differential Revision: D2987288

fb-gh-sync-id: f477611
shipit-source-id: f477611