write filenames as ascii only
Summary:public
In Python 2, `shutil.rmtree()`, `os.walk()`, `os.path.join()`, etc. are not
happy when the locale cannot decode the filenames they have to deal with.
Decrease the likelihood of this happening by making the file names generated by
infer ascii-only.

Also ignore character decoding errors optimistically when reading the json
report file.

Add tests that we are able to run the analysis and report the bug on a function
with a utf8 name, and that we are able to remove the previous results
directory.

closes #287

Reviewed By: cristianoc

Differential Revision: D3058858

fb-gh-sync-id: b88cd35
shipit-source-id: b88cd35