Allow relative or absolute names in ondemand_file
Summary:public
Allow the value of the INFER_ONDEMAND_FILE environment variable to be
either an absolute path, or relative to the project root.

Likewise, allow the entries in the file determined by
INFER_ONDEMAND_FILE to be either absolute or relative to the project
root.  Note, however, that if they are absolute but not under the
project root, they will not be found.

Reviewed By: cristianoc

Differential Revision: D3098489

fb-gh-sync-id: e861300
fbshipit-source-id: e861300