Make eradicate work in the absence of procdescs for callees.
Summary:public
Eradicate need the procedure attributes for callees.
It relies on the java front-end to create proc descs for callees that are declared but not defined.
This diff remove that needs, and when a callee without prodedure attributes is found, it creates one on the fly. The attribute created is similar to what the Java front-end would do, except
that the number and types of arguments are part of the call instruction, so they can
be used to create the formal parameters.

Reviewed By: jeremydubreil

Differential Revision: D3073904

fb-gh-sync-id: 381ff67
fbshipit-source-id: 381ff67