add cmake to the list of make-like tools
Summary:public
Simplifies the way to configure cmake to run infer's fake compiler commands.
Instead of `CC=/path/to/infer/infer/lib/capture/clang cmake .`, which is what
is advised on #25 and is now outdated because the paths to infer's clang has
changed since, simply run `infer -- cmake .`. The only caveat is that infer
tries to analyze the end result, and prints "No issues found". This could be
fixed later.

Reviewed By: jeremydubreil

Differential Revision: D3093162

fb-gh-sync-id: 99df50a
fbshipit-source-id: 99df50a