Fix control-flow graph shape when conditional is last instruction.
Summary:public

When a conditional is the last instruction, there will be a join node leading directly to the exit node.
Some instructions, such as nullification of dead variables, and abstraction, are added to the control flow graph automatically. But, join nodes cannot contain instructions. So when a procedure ends with a conditional, there might be no place to store these instructions.
This diff adds one extra node between the join and the exit node in that situation.

Reviewed By: jvillard

Differential Revision: D3179056

fb-gh-sync-id: 2b9cd7e
fbshipit-source-id: 2b9cd7e