Do not report @NoAllocation and @PerformanceCritical violations in "unlikely" branches
Summary: For performance critical sections of the code, this checker detects memory allocations or calls to methods annotated as expensive. However, such cases of memory allocations or expensive calls are acceptable is occuring in rare cases. This diff adds supports for the "unlikely" branch prediction method and does not track expensive calls in unlikely branches.

Reviewed By: sblackshear

Differential Revision: D3193473

fb-gh-sync-id: ea87e49
fbshipit-source-id: ea87e49