add matchers from inferconfig to the CLI
Summary:
Part of the migration of .inferconfig-specific options into options accepted
both by .inferconfig and the CLI.

This changes the behaviour of Infer in that we now create matchers eagerly
instead of lazily. I think it's ok because I suspect what's really important is
not laziness but memoisation, and thus laziness was just an implementation
detail. If I'm wrong please yell, it should be easy to revert to a lazy
behaviour if really needed.

Reviewed By: jberdine

Differential Revision: D3304792

fbshipit-source-id: 1ddde6d