Modify model of std::vector
Summary:
Turns out, analyzer was getting confused with complicated
model and it was reporting empty access in places it
shouldn't. Fixing backend is not trivial (tracing mode is the answer),
but the model can be simplified.
It introduces the problem that get() method doesn't return fresh value
every time, but we should be able to change backend later to deal with it.

Reviewed By: sblackshear

Differential Revision: D3328228

fbshipit-source-id: dddbaf8