Remove specs before rebuilding models
Summary:
Specs files can become stale and cause crashes if the serialized format
changes.  It is unclear how to add the right dependencies to the
existing targets, so this patch just removes them.  Perhaps when the
specs are built we should drop an empty file to time-stamp them, and add
a target that depends on e.g. InferAnalyze and produces that file, and
then add dependencies on it.

Reviewed By: jvillard

Differential Revision: D3358460

fbshipit-source-id: d8aee48