Fix C++ models of c library
Summary:
When analyzing C model in C++, we were seeing some SKIP function triggered by generated constructors/operators= for C structs.
In C they weren't present, but in C++ compiler generates them for us. To avoid this (and future) problems
with models, translate all functions that are needed when computing the model

Reviewed By: dulmarod

Differential Revision: D3561873

fbshipit-source-id: f8ad2a0