Move comments to improve attachment
Summary:
Cosmetic changes to comments to improve the results of the Reason
comment attachment logic.

These were found using `git grep -nH -e 'in[ ]*(\*'` although the
attachment logic seems ok if the associated `let` is on the same line.

Some others were found with `git grep -nH -e ')[ ]*(\*'` although the
attachment logic seems ok if the associated `(` is on the same line.

Reviewed By: jvillard

Differential Revision: D3654027

fbshipit-source-id: 122aa3b