Add module Core in Prop to make transitions between exposed and normalized explicit.
Summary:
Add module `Core` in Prop to contain the implementation of a prop as a record. The record is private so that pattern matching is unchanged.
Added new function to set individual fields, enforcing that the type becomes exposed when constructing explicitly.
Added function unsafe_cast_to_normal to mark all the cases where a coercion to a normalized type is used.
This is purely a refactoring diff that only affects types, it should have no runtime consequence.

Reviewed By: jberdine

Differential Revision: D3691342

fbshipit-source-id: fa06e29