do not print infer.py usage message on wrong arguments
Summary:
On wrong arguments (or on no arguments at all), `infer` would spew the error
message of `infer.py`, which makes no sense. Make the python code swallow error
messages and exit with a special code on errors coming from command line
parsing so that the OCaml side is in charge of printing usage messages.

Reviewed By: cristianoc

Differential Revision: D3731594

fbshipit-source-id: fe49cda