Convert Eradicate and Checkers tests to direct format.
Summary:
This diff converts the Eradicate and Checkers tests to the new direct test format, which does not rely on buck or junit.
A self-contained Makefile is used to compile and analyze the test files, including all the dependencies, and a special option in InferPrint is used to produce a file of expected results `issues.exp`, which is checked into the repository.
Having an explicit Makefile makes it easy to edit and compile one set of test files in isolation, to investigate test failures, do debugging, etc.
A bunch of boilerplate code is removed. For example,  the single file of expected results `issues.exp` replaces the 1.5K LOC in `endtoend/java/eradicate`.

Reviewed By: jvillard

Differential Revision: D3764632

fbshipit-source-id: 6c68ab8