[clang] rewrite plugin attachment logic in OCaml
Summary:
This changes executions of the former InferClang into a function call. In
particular, it can be called several times per execution.

The new InferClang must be called as if it was clang, and knows how to run
clang with our plugin to get the AST of the source file.

Reviewed By: akotulski

Differential Revision: D3981017

fbshipit-source-id: 7af6490