[backend] Perf fix: only load tenv and cfg files once per each file, instead of once per procedure.
Summary:
This fixes a perf issue on large files, where a copy of the type environment and control flow graph were loaded for each procedure analyzed in the file.
If the type environment or the control flow graph are big, and the file contains many procedures, this can cause a big memory overhead.

Reviewed By: jvillard

Differential Revision: D4008655

fbshipit-source-id: 11d07c1