[clang] translate global var initializers
Summary:
Create dummy functions representing the initializers of global variables. This
is so we can implement checks in the backend that can look at the initializer
expressions of global variables. We try not to create these dummy functions
when the initializer is not present, although for some reason we sometimes end
up with empty initializers.

Also add source file info to global variables in the backend (Pvar.re).

Reviewed By: sblackshear

Differential Revision: D3780238

fbshipit-source-id: 2dca87e