Fix issue with timestamps leading to files not analyzed in -reactive mode when the capture takes less than 1 second
Summary:
Fix an issue where, when `-reactive` mode is used, files captured in the first second are not considered modified, and are not analyzed. This happens because file timestamps are used, and the resolution is one second.
Change the front-ends to change the timestamp of the directory where artifacts are created, so that the timestamps are 1 second in the future.

Small reactive commands such as the following now analyze correctly:
  rm -rf infer-out && infer --reactive -- clang -c test.c

Reviewed By: jberdine

Differential Revision: D4050689

fbshipit-source-id: 6271860