[quandary] propagating taint from unknown procedures and constructors
Summary:
Right now, taint gets lost if it flows into a constructor or procedure whose implementation is missing.
Since the core Java (e.g., String) and Android classes (e.g, Intent) are among these, this is bad.
We could handle this by writing a bunch of models instead, but that would be a lot of work (plus we may still miss cases).

Reviewed By: jvillard

Differential Revision: D4051591

fbshipit-source-id: 65851c8