[infer][java] Store the procedure descriptions in the summary to run the analyis with lazy dynamic dispatch across modules
Summary: The lazy dynamic dispatch algorithm works by re-analyzing the generic methods with the more specialized types encountered during the symbolic execution. In order to do that, the analysis must access the procedure description of the method to reanalyze in order to run the analysis of the specialized procedure description on demand. This diff adds the procedure description on the summary as the summary are stored in the Buck cache and can easily be retrieved by procname.

Reviewed By: sblackshear

Differential Revision: D4077415

fbshipit-source-id: c2f1cc8