[quandary] support `StringBuilder`'s and other methods for propagating `String` taint
Summary:
Our default strategy for handling unknown code is to propagate taint from the actuals to the return value.
But for commonly-used methods like `StringBuilder.append` (used every time you do `+` with a string in Java), this doesn't work.
The taint should be propagated to both the receiver and the return value in these cases.

I'm considering a solution where we always propagate taint to the receiver of unknown functions in the future, but I am concerned about the performance.
So let's stick with a few special string cases for now.

Reviewed By: cristianoc

Differential Revision: D4124355

fbshipit-source-id: 5b2a232